# backend-innotech/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "InnoTech Preinscription API",
        "description": "Documentation for the InnoTech University APIS ",
        "contact": {
            "name": "Université InnoTech",
            "email": "admissions@InnoTech-university.tn"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/professors": {
            "get": {
                "tags": [
                    "Professors"
                ],
                "summary": "List all professors",
                "operationId": "fd97c52f555187931df0534dda2d688e",
                "responses": {
                    "200": {
                        "description": "List retrieved successfully"
                    }
                }
            },
            "post": {
                "tags": [
                    "Professors"
                ],
                "summary": "Create a new professor",
                "operationId": "7a2d185458c37adbd0353ab421eeb62a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "department_id",
                                    "translations"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "photo_url": {
                                        "type": "string"
                                    },
                                    "cv_attached_file": {
                                        "type": "string"
                                    },
                                    "department_id": {
                                        "type": "integer"
                                    },
                                    "translations": {
                                        "properties": {
                                            "fr": {
                                                "required": [
                                                    "first_name",
                                                    "last_name"
                                                ],
                                                "properties": {
                                                    "first_name": {
                                                        "type": "string"
                                                    },
                                                    "last_name": {
                                                        "type": "string"
                                                    },
                                                    "speciality": {
                                                        "type": "string"
                                                    },
                                                    "grade": {
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object"
                                            },
                                            "en": {
                                                "properties": {
                                                    "first_name": {
                                                        "type": "string"
                                                    },
                                                    "last_name": {
                                                        "type": "string"
                                                    },
                                                    "speciality": {
                                                        "type": "string"
                                                    },
                                                    "grade": {
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object"
                                            },
                                            "ar": {
                                                "properties": {
                                                    "first_name": {
                                                        "type": "string"
                                                    },
                                                    "last_name": {
                                                        "type": "string"
                                                    },
                                                    "speciality": {
                                                        "type": "string"
                                                    },
                                                    "grade": {
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Professor created successfully"
                    }
                }
            }
        },
        "/api/professors/{id}": {
            "get": {
                "tags": [
                    "Professors"
                ],
                "summary": "Get a professor by ID",
                "operationId": "dcad8026df3c4269952b9da3b699d9a8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Professor retrieved successfully"
                    }
                }
            },
            "put": {
                "tags": [
                    "Professors"
                ],
                "summary": "Update a professor",
                "description": "Updates a professor by ID",
                "operationId": "updateProfessor",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Professor ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "first_name",
                                    "last_name",
                                    "email",
                                    "speciality",
                                    "grade"
                                ],
                                "properties": {
                                    "first_name": {
                                        "type": "string",
                                        "example": "Rami"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "example": "Ben Yedder"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "professor@InnoTech.tn"
                                    },
                                    "speciality": {
                                        "type": "string",
                                        "example": "AI & Data Science"
                                    },
                                    "grade": {
                                        "type": "string",
                                        "example": "Assistant"
                                    },
                                    "department_id": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Professor updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Updated successfully"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Professor not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Professors"
                ],
                "summary": "Delete a professor",
                "operationId": "5172e5f8ee956a691c4ba5c1edaf7262",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Professor deleted successfully"
                    }
                }
            }
        },
        "/api/departments/{id}/professors": {
            "get": {
                "tags": [
                    "Professors"
                ],
                "summary": "List all professors in a specific department",
                "operationId": "b4299a76bd2f8eb3d4fc140e405ab642",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Professors retrieved successfully"
                    }
                }
            }
        },
        "/api/professors/slug/{slug}": {
            "get": {
                "tags": [
                    "Professors"
                ],
                "summary": "Get a professor by profile slug",
                "operationId": "f15efd817f2f65d4760eead14523c171",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Slug of the professor",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Professor retrieved successfully"
                    },
                    "404": {
                        "description": "Professor not found"
                    }
                }
            }
        },
...
    },
...
}